pyproject.toml
[project]
name = "project1"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11, <3.13"
dependencies = [
    "numpy (>=2.4.1,<3.0.0)",
    "tensorflow (>=2.20.0,<3.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
]

[tool.poetry]
packages = [{include = "project1", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
